remove ppx_refinements description from the dune-project
it seems like we only need to mention about the libraries inside our
project in the dune-project if we ever want to make these libraries
public (by public I mean giving outside world the ability to install
just ppx_refinements library), but we intend to use it as an internal
library. Therefore we don't need to talk about it in dune-project. Same
for the type-checker. The type checker that we'll write is intended to
be part of the the refinement_project project that we're building thus
we don't need a section for it in dune-project. Accordingly, it won't
have a public_name field in its dune file in type_checker/dune.